face-scoring/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
insightface==0.7.3
onnxruntime==1.19.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
httpx==0.27.0
